parcial completo

+++ src/main/java/com/uniminuto/clinica/entity/Medico.java
@@ -24,10 +24,10 @@ public class Medico implements Serializable {
      */
     private static final long serialVersionUID = 1L;
     
-       @Id
+    @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id; 
+    private Integer id;
     
     @Column(name = "tipo_documento")
     private String tipoDocumento;
